bugfix merge to develop

+++ ethics-g4-app/client/src/pages/AdminDashboard.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import StyledDashboard from "../styled/Dashboard.styled";
+import StyledAdminDashboard from "../styled/AdminDashboard.styled";
 
 const AdminDashboard = () => {
   const [users, setUsers] = useState([]);
@@ -45,21 +45,26 @@ const AdminDashboard = () => {
   );
 
   const deleteUser = (userId) => {
-    fetch(`${import.meta.env.VITE_SERVER_URL}/api/users/${userId}/delete/`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
-      .then((response) => {
-        if (!response.ok) {
-          throw new Error("Failed to delete user");
-        }
-        // Refresh the users list or handle UI update
-        setFetchTrigger((prev) => prev + 1);
+    // Display confirmation dialog before deleting user
+    const confirmDelete = window.confirm("Are you sure you want to delete this user?");
+    if (confirmDelete) {
+      fetch(`${import.meta.env.VITE_SERVER_URL}/api/users/${userId}/delete/`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
       })
-      .catch((error) => console.error("Error deleting user:", error));
+        .then((response) => {
+          if (!response.ok) {
+            throw new Error("Failed to delete user");
+          }
+          // Refresh the users list or handle UI update
+          setFetchTrigger((prev) => prev + 1);
+        })
+        .catch((error) => console.error("Error deleting user:", error));
+    }
   };
+  
 
   // Function to update user role
   const updateUserRole = (userId, newRole, departmentCode) => {
@@ -84,6 +89,10 @@ const AdminDashboard = () => {
 
   const addDepartment = async () => {
     try {
+      if (!department_name || !department_code) {
+        throw new Error("Department name and code are required");
+      }
+  
       const response = await fetch(
         `${import.meta.env.VITE_SERVER_URL}/api/departments/add`,
         {
@@ -97,7 +106,7 @@ const AdminDashboard = () => {
           }),
         }
       );
-
+  
       if (!response.ok) {
         throw new Error("Failed to add department");
       }
@@ -105,24 +114,29 @@ const AdminDashboard = () => {
       setDepartmentName("");
       setDepartmentCode("");
     } catch (error) {
-      console.error("Error deleting department:", error.message);
+      console.error("Error adding department:", error.message);
     }
   };
+  
 
   const deleteDepartment = async (departmentId) => {
-    try {
-      const response = await fetch(
-        `${import.meta.env.VITE_SERVER_URL}/api/departments/${departmentId}`,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        }
-      );
-      setFetchTrigger((prev) => prev + 1);
-    } catch (error) {
-      console.error("Error deleting department:", error.message);
+    // Display confirmation dialog before deleting department
+    const confirmDelete = window.confirm("Are you sure you want to delete this department?");
+    if (confirmDelete) {
+      try {
+        const response = await fetch(
+          `${import.meta.env.VITE_SERVER_URL}/api/departments/${departmentId}`,
+          {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+            },
+          }
+        );
+        setFetchTrigger((prev) => prev + 1);
+      } catch (error) {
+        console.error("Error deleting department:", error.message);
+      }
     }
   };
 
@@ -131,17 +145,26 @@ const AdminDashboard = () => {
     updateUserRole(selectedUserId, "admin", departmentCode);
   };
 
-  const renderDepartmentButtons = () => {
-    return departments.map((department) => (
-      <button
-        key={department.id}
-        className="btn"
-        onClick={() => handleDepartmentButtonClick(department.code)}
-      >
-        {department.name}
-      </button>
-    ));
-  };
+ const renderDepartmentButtons = () => {
+   const adminDepartments = users.reduce((departments, user) => {
+     if (user.role === "admin" && user.admin_of_department) {
+       departments.add(user.admin_of_department);
+     }
+     return departments;
+   }, new Set());
+
+   return departments.map((department) => (
+     <button
+       key={department.id}
+       className="btn"
+       disabled={adminDepartments.has(department.code)}
+       onClick={() => handleDepartmentButtonClick(department.code)}
+     >
+       {department.name}
+     </button>
+   ));
+ };
+
 
   const updateAdminRole = (user_id) => {
     setSelectedUserId((prevUserId) =>
@@ -150,119 +173,155 @@ const AdminDashboard = () => {
   };
 
   return (
-    <StyledDashboard>
-      <div className="header">
-      <Link to={`${import.meta.env.VITE_SERVER_URL}/api/users`}>
-        <h1>Admin Dashboard</h1>
-      </Link>
+    <StyledAdminDashboard>
       <div>
-        <button className="btn" onClick={() => setCurrentView("users")}>
-          Manage Users
-        </button>
-        <button className="btn" onClick={() => setCurrentView("departments")}>
-          Manage Departments
-        </button>
-      </div>
-      {currentView === "users" && (
-        <>
-          <div>
+        <Link to={`${import.meta.env.VITE_SERVER_URL}/api/users`}>
+          <h1>Admin Dashboard</h1>
+        </Link>
+        <div className="tabs">
+          <button
+            title="Switch tabs"
+            className="btn"
+            onClick={() => setCurrentView("users")}
+          >
+            Manage Users
+          </button>
+          <button
+            title="Switch tabs"
+            className="btn"
+            onClick={() => setCurrentView("departments")}
+          >
+            Manage Departments
+          </button>
+        </div>
+        {currentView === "users" && (
+          <>
             <input
               type="text"
               placeholder="Search by username"
               value={searchTerm}
               onChange={(e) => setSearchTerm(e.target.value)}
             />
-            <ul>
-              {filteredUsers.map((user) => (
-                <li key={user.user_id}>
-                  <div>
-                    <img
-                      src={user.img}
-                      alt={user.username}
-                      referrerPolicy="no-referrer"
-                    />
-                    <p>{user.username}</p>
-                    <p>Role: {user.role}</p>
-                    {user?.admin_of_department && (
-                      <p>Department: {user.admin_of_department}</p>
+            <div>
+              <div className="header">
+                <p className="username">Username</p>
+                <p className="role">Role</p>
+                <p className="department">Department</p>
+                <p className="actions">Actions</p>
+                <p className="actions"></p>
+              </div>
+            </div>
+            <table>
+              <tbody>
+                {filteredUsers.map((user) => (
+                  <>
+                    <tr key={user.user_id}>
+                      <td className="username">
+                        <img
+                          title={"User ID: " +user.user_id}
+                          src={user.img}
+                          alt={user.username}
+                          referrerPolicy="no-referrer"
+                        />
+                        {user.username}
+                      </td>
+                      <td className="role">{user.role}</td>
+                      <td className="department">{user.admin_of_department}</td>
+                      <td className="actions">
+                        <button
+                          className="btn"
+                          onClick={() => updateAdminRole(user.user_id)}
+                        >
+                          Make Admin
+                        </button>
+                        <button
+                          className="btn"
+                          disabled={selectedUserId === user.user_id}
+                          onClick={() =>
+                            updateUserRole(user.user_id, "staff", "")
+                          }
+                        >
+                          Make Staff
+                        </button>
+                        <button
+                          className="btn"
+                          disabled={selectedUserId === user.user_id}
+                          onClick={() =>
+                            updateUserRole(user.user_id, "student", "")
+                          }
+                        >
+                          Make Student
+                        </button>
+                        <button
+                          className="btn btn_delete"
+                          disabled={selectedUserId === user.user_id}
+                          onClick={() => deleteUser(user.user_id)}
+                        >
+                          Delete User
+                        </button>
+                      </td>
+                    </tr>
+                    {selectedUserId === user.user_id && (
+                      <tr key={selectedUserId}>
+                        {selectedUserId !== null &&
+                          selectedUserId === user.user_id &&
+                          renderDepartmentButtons()}
+                      </tr>
                     )}
-                    <button
-                      className="btn"
-                      onClick={() => updateAdminRole(user.user_id)}
-                    >
-                      Change to Admin
-                    </button>
-                    <button
-                      className="btn"
-                      disabled={selectedUserId === user.user_id}
-                      onClick={() => updateUserRole(user.user_id, "staff", "")}
-                    >
-                      Change to Staff
-                    </button>
-                    <button
-                      className="btn"
-                      disabled={selectedUserId === user.user_id}
-                      onClick={() =>
-                        updateUserRole(user.user_id, "student", "")
-                      }
-                    >
-                      Change to User
-                    </button>
-                    <button
-                      className="btn"
-                      disabled={selectedUserId === user.user_id}
-                      onClick={() => deleteUser(user.user_id)}
-                    >
-                      Delete User
-                    </button>
-                  </div>
-                  {selectedUserId !== null &&
-                    selectedUserId === user.user_id &&
-                    renderDepartmentButtons()}
-                </li>
-              ))}
-            </ul>
-          </div>
-        </>
-      )}
-      {currentView === "departments" && (
-        <>
-          <h1> Departments</h1>
-          <input
-            type="text"
-            placeholder="Department name"
-            value={department_name}
-            onChange={(e) => setDepartmentName(e.target.value)}
-          />
-          <input
-            type="text"
-            placeholder="Department code"
-            value={department_code}
-            onChange={(e) => setDepartmentCode(e.target.value)}
-          />
-          <button className="btn" onClick={() => addDepartment()}>
-            Create a Department
-          </button>
-          <ul>
-            {filteredDepartments.map((department) => (
-              <li key={department.id}>
-                <div>
-                  <p>{department.name}</p>
-                  <p>Code: {department.code}</p>
-                  <button
-                    className="btn"
-                    onClick={() => deleteDepartment(department.id)}
-                  >
-                    Delete
-                  </button>
-                </div>
-              </li>
-            ))}
-          </ul>
-        </>
-      )}
-    </div>
-   </StyledDashboard>
+                  </>
+                ))}
+              </tbody>
+            </table>
+          </>
+        )}
+        {currentView === "departments" && (
+          <>
+            <div className="create-departments">
+              <input
+                type="text"
+                placeholder="Department name"
+                value={department_name}
+                onChange={(e) => setDepartmentName(e.target.value)}
+              />
+              <input
+                type="text"
+                placeholder="Department code"
+                value={department_code}
+                onChange={(e) => setDepartmentCode(e.target.value)}
+              />
+              <button className="btn btn_appro" onClick={() => addDepartment()}>
+                Create Department
+              </button>
+            </div>
+            <table className="departments-table">
+              <thead>
+                <tr>
+                  <th>Department Name</th>
+                  <th>Department Code</th>
+                  <th>Actions</th>
+                </tr>
+              </thead>
+              <tbody>
+                {filteredDepartments.map((department) => (
+                  <tr key={department.id}>
+                    <td>{department.name}</td>
+                    <td>{department.code}</td>
+                    <td>
+                      <button
+                        className="btn btn_delete"
+                        onClick={() => deleteDepartment(department.id)}
+                      >
+                        Delete
+                      </button>
+                    </td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
+          </>
+        )}
+      </div>
+    </StyledAdminDashboard>
   );
 };
